Fix compiler warnings. Cache flush not working

+++ src/Reverend/object/tile.h
@@ -1,4 +1,5 @@
 #pragma once
+
 #ifndef TILE_H
 #define TILE_H
 
@@ -11,8 +12,8 @@ public:
 	virtual void draw();
 	virtual void update();
 	virtual void clean();
-	virtual int getX() { return position_.getX(); }
-	virtual int getY() { return position_.getX(); }
+	virtual int getX() { return (int)position_.getX(); }
+	virtual int getY() { return (int)position_.getX(); }
 private:
 	Vector2d position_;
 };
